refactor: ha_cluster_info: split to several modules
Implementing changes proposed in code review
