+++ .sanity-ansible-ignore-2.13.txt
@@ -23,4 +23,8 @@ plugins/modules/pcs_qdevice_certs.py import-3.8!skip
 plugins/modules/ha_cluster_info.py compile-2.7!skip
 plugins/modules/ha_cluster_info.py import-2.7!skip
 plugins/modules/ha_cluster_info.py validate-modules:missing-gplv3-license
+plugins/module_utils/ha_cluster_lsr/info/exporter.py compile-2.7!skip
+plugins/module_utils/ha_cluster_lsr/info/exporter.py import-2.7!skip
+plugins/module_utils/ha_cluster_lsr/info/loader.py compile-2.7!skip
+plugins/module_utils/ha_cluster_lsr/info/loader.py import-2.7!skip
 tests/ha_cluster/unit/test_ha_cluster_info.py shebang!skip
